+++ payment-service/src/main/java/com/palja/payment_service/application/type/PaymentRetryAction.java
@@ -1,4 +1,4 @@
-package com.palja.payment_service.application;
+package com.palja.payment_service.application.type;
 
 public enum PaymentRetryAction {
     COMPLETE,
